Stopped using texture atlases for the tilemap.
The automatic texture atlas importer was putting textures and their normal maps in different parts of the texture map so the tiled importer was receiving the wrong texture for each index. It might be possible to create my own texture atlas programatically with textures in the correct order. For now it works ok but it may be neccesary for mobile performance further down the line.